+++ Changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.1.0] - 2025-05-02
+- Get subjects, sender, content messages from gamil API,
+- Set ranges dates for messages
+
 ## [1.0.0] - 2025-04-15
 - Initial release
 - Integrated DynamoDB and Gmail APIs